# splinterd/api/static/openapi.yml
...
openapi: '3.0.0'
info:
  version: 0.3.12
  title: SplinterD API
  description: REST API for SplinterD
servers:
  - url: http://localhost:9000/api
paths:
...
  /admin/nodes:
    post:
      tags:
        - Node Registry
      description: Add node to the Node Registry
      parameters:
        - $ref: "#/components/parameters/protocol_version"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisteredNode'
      responses:
        200:
          description: The node was added to the registry
        400:
          description: The request was malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: The node already exists or is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Node Registry
      description: List nodes in the Node Registry
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: offset
          in: query
          description: paging offset
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: maximum number of items to return (max 100)
          required: false
          schema:
            type: integer
            default: 100
        - name: filter
          in: query
          description: |
            url-encodeded stringified JSON containing property filters on the
            node's metadata properties in the format
              {METADATA_PROPERTY:[{"operator":OPERATOR,"value":VALUE}]}
          required: false
          schema:
            type: string
          example: "%7B%22company%22%3A%5B%22%3D%22%2C%22Cargill%22%5D%7D"
      responses:
        200:
          description: list of nodes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RegisteredNode'
                  paging:
                    $ref: '#/components/schemas/Paging'
        400:
          description: Request was malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/nodes/{identity}:
    get:
      tags:
        - Node Registry
      description: Fetch a nodes in the Node Registry by their identity
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: identity
          in: path
          description: identity of node to fetch
          required: true
          schema:
            type: string
      responses:
        200:
          description: Node
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/RegisteredNode"
        404:
          description: The node with {identity} was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Node Registry
      description: Add a node to or replace a node in the Node Registry. This action is idempotent.
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: identity
          in: path
          description: identity of node to add/replace
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisteredNode'
      responses:
        200:
          description: The node in the registry has been replaced
        400:
          description: The request was malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: The node was invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Node Registry
      description: Delete a node from the Node Registry
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: identity
          in: path
          description: identity of node to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: The node has been deleted from the registry
        404:
          description: The node with {identity} was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    protocol_version:
      schema:
        type: integer
        example: 1
      description: The protocol version which the client can understand
      in: header
      name: SplinterProtocolVersion
  schemas:
    Error:
      additionalProperties: false
      properties:
        message:
          description: A message describing the error that occurred
          type: string
          example: DatabaseError({description})
      required:
        - message
...
    RegisteredNode:
      type: object
      properties:
        identity:
          type: string
        endpoint:
          type: string
        display_name:
          type: string
        metadata:
          type: object
      example:
        identity: node-123123-asdf
        endpoint: tls://12.0.0.123:8431
        display_name: Cargill - Node 1
        metadata:
          company: Cargill
          status: Up
...
    Paging:
      type: object
      properties:
        current:
          type: string
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        first:
          type: string
        prev:
          type: string
        next:
          type: string
        last:
          type: string
      example:
        current: /admin/nodes?offset=10&limit=10
        offset: 10
        limit: 10
        total: 50
        first: /admin/nodes?offset=0&limit=10
        prev: /admin/nodes?offset=0&limit=10
        next: /admin/nodes?offset=20&limit=10
        last: /admin/nodes?offset=40&limit=10
...